refactor(query-runner): rename file
